Enhance ADK vs Go Micro comparison and apply lint fixes (#2994)
* docs: compare Go Micro with Google ADK in the comparison guide

Adds a 'vs Agent Frameworks (Google ADK)' section: ADK builds an agent,
Go Micro builds the distributed system the agent lives in (agents are
services in the mesh). Covers the category difference, a feature table,
when to choose each, and MCP/A2A interoperability.


* docs: replace ADK comparison slogan with concrete explanation

State plainly what each tool provides (ADK builds an agent process; Go Micro
builds the surrounding service mesh) instead of marketing phrasing.

* lint: apply golangci-lint autofixes; exclude ST1003 and demo errcheck

Mechanical, behaviour-preserving fixes applied by 'golangci-lint run --fix':
gofmt, misspell (US spelling), usestdlibvars (http.Method*/Status*), unconvert,
and the auto-fixable staticcheck simplifications (QF*, S1017/S1019/S1023/S1039).

Config: exclude ST1003 (remaining offenders are exported API renames, e.g.
web.Id, which would break compatibility) and skip errcheck for examples/ and
internal/harness/ (demo code where fire-and-forget is intentional).

Build and test compilation verified.


* lint: WIP cleanup checkpoint (errcheck config + partial fixes)

Checkpoint of an in-progress golangci-lint cleanup (background pass). Builds
cleanly; lint is not yet zero. Follow-up commit will complete the cleanup and
switch CI to a blocking full-tree lint.


---------

+++ codec/grpc/util.go
@@ -20,7 +20,7 @@ func decode(r io.Reader) (uint8, []byte, error) {
 	}
 
 	// get encoding format e.g compressed
-	cf := uint8(header[0])
+	cf := header[0]
 
 	// get message length
 	length := binary.BigEndian.Uint32(header[1:])
@@ -54,7 +54,7 @@ func encode(cf uint8, buf []byte, w io.Writer) error {
 	header := make([]byte, 5)
 
 	// set compression
-	header[0] = byte(cf)
+	header[0] = cf
 
 	// write length as header
 	binary.BigEndian.PutUint32(header[1:], uint32(len(buf)))
